Sanitize OpenRouter keys, replace hardcoded secrets with env usage; docs: update guides to placeholders; scripts: remove default key. Also include AI moderation additions and related backend/frontend changes. No secrets committed.

+++ webApps/client/src/tools/wavTools/wav_recorder.ts
@@ -133,7 +133,7 @@ export class WavRecorder {
         throw new Error(`Minimum "fromSampleRate" is 3000 (3kHz)`);
       }
       if (!data) {
-        data = WavPacker.floatTo16BitPCM(float32Array) as Int16Array;
+        data = WavPacker.floatTo16BitPCM(float32Array);
       }
       const audio = {
         bitsPerSample: 16,
